+++ constants.js
@@ -3,15 +3,21 @@ var p    = require('path');
 var fs   = require('fs');
 var util = require('util');
 
-var HOME = process.env.PM2_HOME || process.env.HOME || process.env.HOMEPATH;
+/**
+ * Handle PM2 root folder relocation
+ */
+var PM2_ROOT_PATH = '';
 
-var DEFAULT_FILE_PATH = p.resolve(HOME, '.pm2');
+if (process.env.PM2_HOME)
+  PM2_ROOT_PATH = process.env.PM2_HOME;
+else
+  PM2_ROOT_PATH = p.resolve(process.env.HOME || process.env.HOMEPATH, '.pm2');
 
 /**
  * Constants variables used by PM2
  */
 var csts = {
-  PM2_CONF_FILE : p.join(DEFAULT_FILE_PATH, 'conf.js'),
+  PM2_CONF_FILE          : p.join(PM2_ROOT_PATH, 'conf.js'),
 
   CODE_UNCAUGHTEXCEPTION : 100,
   CONCURRENT_ACTIONS     : 1,
@@ -51,16 +57,16 @@ var csts = {
  * Defaults variables
  */
 var default_conf = {
-  DEFAULT_FILE_PATH        : DEFAULT_FILE_PATH,
-  PM2_LOG_FILE_PATH        : p.join(DEFAULT_FILE_PATH, 'pm2.log'),
-  PM2_PID_FILE_PATH        : p.join(DEFAULT_FILE_PATH, 'pm2.pid'),
-  DEFAULT_PID_PATH         : p.join(DEFAULT_FILE_PATH, 'pids'),
-  DEFAULT_LOG_PATH         : p.join(DEFAULT_FILE_PATH, 'logs'),
-  DUMP_FILE_PATH           : p.join(DEFAULT_FILE_PATH, 'dump.pm2'),
+  PM2_ROOT_PATH        : PM2_ROOT_PATH,
+  PM2_LOG_FILE_PATH        : p.join(PM2_ROOT_PATH, 'pm2.log'),
+  PM2_PID_FILE_PATH        : p.join(PM2_ROOT_PATH, 'pm2.pid'),
+  DEFAULT_PID_PATH         : p.join(PM2_ROOT_PATH, 'pids'),
+  DEFAULT_LOG_PATH         : p.join(PM2_ROOT_PATH, 'logs'),
+  DUMP_FILE_PATH           : p.join(PM2_ROOT_PATH, 'dump.pm2'),
 
-  DAEMON_RPC_PORT          : p.join(DEFAULT_FILE_PATH, 'rpc.sock'),
-  DAEMON_PUB_PORT          : p.join(DEFAULT_FILE_PATH, 'pub.sock'),
-  INTERACTOR_RPC_PORT      : p.join(DEFAULT_FILE_PATH, 'interactor.sock'),
+  DAEMON_RPC_PORT          : p.join(PM2_ROOT_PATH, 'rpc.sock'),
+  DAEMON_PUB_PORT          : p.join(PM2_ROOT_PATH, 'pub.sock'),
+  INTERACTOR_RPC_PORT      : p.join(PM2_ROOT_PATH, 'interactor.sock'),
 
   GRACEFUL_TIMEOUT         : parseInt(process.env.PM2_GRACEFUL_TIMEOUT) || 8000,
   GRACEFUL_LISTEN_TIMEOUT  : parseInt(process.env.PM2_GRACEFUL_LISTEN_TIMEOUT) || 4000,
@@ -69,9 +75,9 @@ var default_conf = {
   WEB_INTERFACE            : parseInt(process.env.PM2_API_PORT)  || 9615,
   MODIFY_REQUIRE           : process.env.PM2_MODIFY_REQUIRE || false,
 
-  INTERACTOR_LOG_FILE_PATH : p.join(DEFAULT_FILE_PATH, 'agent.log'),
-  INTERACTOR_PID_PATH      : p.join(DEFAULT_FILE_PATH, 'agent.pid'),
-  INTERACTION_CONF         : p.join(DEFAULT_FILE_PATH, 'agent.json')
+  INTERACTOR_LOG_FILE_PATH : p.join(PM2_ROOT_PATH, 'agent.log'),
+  INTERACTOR_PID_PATH      : p.join(PM2_ROOT_PATH, 'agent.pid'),
+  INTERACTION_CONF         : p.join(PM2_ROOT_PATH, 'agent.json')
 };
 
 /**
@@ -80,7 +86,7 @@ var default_conf = {
 
 if (fs.existsSync(csts.PM2_CONF_FILE)) {
   try {
-    var extra = require(csts.PM2_CONF_FILE)(DEFAULT_FILE_PATH);
+    var extra = require(csts.PM2_CONF_FILE)(PM2_ROOT_PATH);
     default_conf = util._extend(default_conf, extra);
   } catch(e) {
     console.error(e.stack || e);